Add UserAPI
