+++ python_anywhere/accounts/serializers.py
@@ -1,0 +1,11 @@
+from rest_framework.serializers import ModelSerializer
+
+from python_anywhere.accounts.models import User
+
+
+class UserSerializer(ModelSerializer):
+    """Serializer for User instances"""
+
+    class Meta:
+        model = User
+        fields = ['username', 'email']
