+++ python_anywhere/conftest.py
@@ -1,4 +1,5 @@
 import pytest
+from dj_rest_auth.utils import jwt_encode
 from faker import Faker
 from rest_framework.test import APIClient
 
@@ -6,12 +7,10 @@ from python_anywhere.accounts.factories import UserFactory
 
 
 @pytest.fixture()
-def logged_in_client(db):
+def logged_in_client_rest(db):
     # login
     password = Faker().password()
     user = UserFactory(password=password)
     client = APIClient()
-    client.login(username=user.email, password=password)
+    client.credentials(HTTP_AUTHORIZATION='Bearer ' + str(jwt_encode(user)[0]))
     yield client, user
-
-
